Commiting the main code of Lepoard Gecko
This code consists the fucntionality to switch the LEDs on-board in
accordance with the Light and Temperature sensors. This incorporates DMA
to collect the values from the Temperature sensor. Employs LEUART to
send the values to atmel in predefined time intervals - defined by
LETIMER0 onduty.
The entire code is in low power mode (E3 and E2)